Fixed lifetime bug in hmac.c
Instead of copying the new hashed key into the key variable (which is
const) I reassign the key pointer to the new buffer containing the
hashed key. But this buffer has a short lifetime which doesn't last for
the rest of the function. For this fix I move the new buffer declaration
so its lifetime lasts long enough.